requirements.txt
requests==2.26.0
django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.10.1
pandas==1.3.5
pycodestyle==2.8.0
flake8==4.0.1
numpy==1.21.4
redis==4.0.2
scipy==1.7.3
statsmodels==0.13.1
lifelines==0.26.4
psycopg2-binary==2.9.2
plotly==5.4.0
autograd==1.3
pycryptodome==3.12.0
filelock
